// src/action/ActionType.enum.ts
export enum ActionType {
  SEND_REQUEST = "SEND_REQUEST",
  REQUEST_RETURNED = "REQUEST_RETURNED",
  REQUEST_FAILED = "REQUEST_FAILED",
  SELECT_TEMPLATE = "SELECT_TEMPLATE",
  ADD_NEW_TEMPLATE = "ADD_NEW_TEMPLATE",
  UPDATE_TEMPLATE = "UPDATE_TEMPLATE",
  START_THE_TRAIN = "START_THE_TRAIN",
  DELETE_TEMPLATE = "DELETE_TEMPLATE",
  CONSUMING_TEMPLATE_START = "CONSUMING_TEMPLATE_START",
  CONSUMING_TEMPLATE_END = "CONSUMING_TEMPLATE_END",
  TEMPLATE_ERROR = "TEMPLATE_ERROR"
}